working on refactoring of the Integration widget
